+++ go/Makefile
@@ -19,7 +19,7 @@ MOCK_SRC_IN =\
     lib/snet/interface.go \
     lib/snet/snetproxy/reconnecter.go \
     lib/snet/snetproxy/io.go \
-    lib/log/log.go
+    lib/log/wrappers.go
 MOCK_SRC_OUT = $(foreach in,$(MOCK_SRC_IN),$(call mock_src_in_to_out,$(in)))
 # This uses the format <stdlib package>/<interface>:
 MOCK_STD_IN = net/Addr
